updated deletion logic for project & analysis

+++ src/api/services/analysis.service.js
@@ -1,11 +1,14 @@
+const BaseService = require('./base.service');
+const ProjectRepository = require('../repository/project.repository');
 const AnalysisRepository = require('../repository/analysis.repository');
+const InteractionRepository = require('../repository/interaction.repository');
 const FlowRepository = require('../repository/flow.repository');
+const MetricsRepository = require('../repository/metrics.repository');
 
-class AnalysisService {
+class AnalysisService extends BaseService {
 
   constructor(req, res) {
-    this.req = req;
-    this.res = res;
+    super(req, res);
   }
 
   async getAnalysisByIdResponse() {
@@ -15,7 +18,7 @@ class AnalysisService {
       const doc = await analysisRepo.getById(id);
       this.res.status(200).json(doc);
     } catch (error) {
-      handlerError(this.res, error);
+      this.handleError(error);
     }
   }
 
@@ -25,7 +28,7 @@ class AnalysisService {
       const doc = await analysisRepo.get(this.req.query);
       this.res.status(200).json(doc);
     } catch (error) {
-      handlerError(this.res, error);
+      this.handleError(error);
     }
   }
 
@@ -37,28 +40,35 @@ class AnalysisService {
       const doc = await analysisRepo.save(analysis);
       this.res.status(200).json(doc);
     } catch (error) {
-      handlerError(this.res, error);
+      this.handleError(error);
     }
   }
 
   async deleteAnalysisResponse() {
     try {
       const id = this.req.swagger.params.id.value;
-      const flowRepo = new FlowRepository();
-      await flowRepo.deleteByAnalysisId(id);
       const analysisRepo = new AnalysisRepository();
-      const doc = await analysisRepo.delete(id);
-      this.res.status(200).json(doc);
+      const analysis = await analysisRepo.getById(id);
+      if (analysis) {
+        const flowRepo = new FlowRepository();
+        await flowRepo.deleteByAnalysisId(id);
+        const interactionRepo = new InteractionRepository();
+        await interactionRepo.deleteByAnalysisId(id);
+        const metricsRepo = new MetricsRepository();
+        await metricsRepo.deleteAnalysisMetrics(id);
+        // TODO: remove/update project metrics if this is the last analysis
+        const projectRepo = new ProjectRepository();
+        await projectRepo.deleteAnalysis(analysis.projectId, analysis);
+        const doc = await analysisRepo.delete(id);
+        this.res.status(200).json(doc);
+      } else {
+        this.res.status(404).json({ message: 'Analysis not found' });
+      }
     } catch (error) {
-      handlerError(this.res, error);
+      this.handleError(error);
     }
   }
 
-}
-
-function handlerError(res, error) {
-  console.log(error);
-  res.status(500).json({ error: "Internal Server Error" });
 }
 
 module.exports = AnalysisService;
